requirements.txt
flask
flask_restful
paho-mqtt
flask_jwt_extended
sqlalchemy
Werkzeug